+++ opentracing_instrumentation/client_hooks/tornado_http.py
@@ -127,9 +127,9 @@ def traced_fetch_impl(real_fetch_impl):
 
         def new_callback(response):
             if hasattr(response, 'code') and response.code:
-                span.add_tag('status_code', '%s' % response.code)
+                span.set_tag('status_code', '%s' % response.code)
             if hasattr(response, 'error') and response.error:
-                span.add_tag('error', '%s' % response.error)
+                span.error('error', '%s' % response.error)
             span.finish()
             return callback(response)