+++ opentracing_instrumentation/http_server.py
@@ -40,19 +40,19 @@ def before_request(request, tracer=None):
         span = tracer.join_trace(operation_name=operation,
                                  parent_trace_context=context)
 
-    span.add_tag('client.http.url', request.full_url)
+    span.set_tag('client.http.url', request.full_url)
 
     remote_ip = request.remote_ip
     if remote_ip:
-        span.add_tag(tags.PEER_HOST_IPV4, remote_ip)
+        span.set_tag(tags.PEER_HOST_IPV4, remote_ip)
 
     caller_name = request.caller_name
     if caller_name:
-        span.add_tag(tags.PEER_SERVICE, caller_name)
+        span.set_tag(tags.PEER_SERVICE, caller_name)
 
     remote_port = request.remote_port
     if remote_port:
-        span.add_tag(tags.PEER_PORT, remote_port)
+        span.set_tag(tags.PEER_PORT, remote_port)
 
     return span