+++ opentracing_instrumentation/local_span.py
@@ -19,7 +19,9 @@
 # THE SOFTWARE.
 from __future__ import absolute_import
 import contextlib2
-from opentracing_instrumentation import get_current_span
+import wrapt
+import tornado.stack_context
+from . import get_current_span, RequestContextManager
 
 
 def func_span(func, tags=None):
@@ -51,3 +53,68 @@ def func_span(func, tags=None):
 
     # TODO convert func to a proper name: module:class.func
     return span.start_child(operation_name=str(func), tags=tags)
+
+
+def traced_function(operation_name, **tags):
+    """
+    A decorator that enables tracing of the wrapped function provided there
+    is a parent span already established. Should not be used with Tornado
+    coroutines because the span is finished as soon as the wrapped function
+    returns a result, which in case of a coroutine will be an incomplete
+    future. For coroutine use `traced_coroutine` decorator instead.
+
+    :param operation_name: used as the name of the child span
+    :param tags: optional tags that can be passed to the new child span
+    :return: returns a tracing decorator
+    """
+
+    @wrapt.decorator
+    def wrapper(wrapped, _, args, kwargs):
+        parent_span = get_current_span()
+        if parent_span is None:
+            return wrapped(*args, **kwargs)
+        with parent_span.start_child(operation_name=operation_name,
+                                     tags=tags) as span:
+            mgr = lambda: RequestContextManager(span)
+            with tornado.stack_context.StackContext(mgr):
+                return wrapped(*args, **kwargs)
+    return wrapper
+
+
+def traced_coroutine(operation_name, **tags):
+    """
+    A decorator that enables tracing of the wrapped Tornado coroutine
+    provided there is a parent span already established.
+
+    :param operation_name: used as the name of the child span
+    :param tags: optional tags that can be passed to the new child span
+    :return: returns a tracing decorator
+    """
+
+    @wrapt.decorator
+    def wrapper(wrapped, _, args, kwargs):
+        parent_span = get_current_span()
+        if parent_span is None:
+            return wrapped(*args, **kwargs)
+
+        span = parent_span.start_child(operation_name=operation_name,
+                                       tags=tags)
+        mgr = lambda: RequestContextManager(span)
+        with tornado.stack_context.StackContext(mgr):
+            try:
+                # cannot use yield inside StackContext, so instead
+                # use a hook on Future completion
+                res = wrapped(*args, **kwargs)
+
+                def done_callback(future):
+                    if future.exception() is not None:
+                        span.error('exception', future.exception())
+                    span.finish()
+
+                res.add_done_callback(done_callback)
+                return res
+            except Exception as e:
+                span.error('exception', e)
+                span.finish()
+                raise
+    return wrapper